Alternate "API Key" and "Bearer Token" in Authorization docs
In the documentation for Authorization header values for the "Dual"
authentication scheme, it is specified that the "API Key" must appear
before the "Bearer Token". However, this is incorrect.

The Gateway API appears to only accept Authorization headers in this
format: `Authorization: ${bearerToken}:${apiKey}:${hmac}`

n.b. It appears that both the Java and JS libs do this correctly. It's
just the documentation that's incorrect.